Fixed bug

+++ app/src/main/java/com/sgc/speedometer/data/model/SpeedometerRecord.kt
@@ -7,7 +7,7 @@ import kotlinx.android.parcel.Parcelize
 data class SpeedometerRecord(
     var duration: Date = Date(0),
     var distance: Double = 0.0,
-    var averageSpeed: Int = 0,
+    var averageSpeed: Double = 0.0,
     var maxSpeed:Double = 0.0,
     var currentSpeed: Double = 0.0
 ) : Parcelable
